3/23--1:33am
Did a bunch of stuff. App launches with stability. GUI for character
creation nearly done.
Changed background image, improved listview with custom layout.

Tried to do do fancy add-on that puts first letter of character on
colored square for each entity, spent about 10 minutes on it, ran into
roadblock, did not pursue as it's not on the documentation. Still looks
neat and could aid lookup speed, as would giving the user ability to
pick a color for the entity's colored square.