3/24 2:32 am
Made a make-shift dynamic array. Turns out arrays in Java can be
declared with variable size and not just with constants.
Also made a custom adapter that does what it's supposed to do (so far)
HOWEVER now the onClick events don't work. I'm sure it's something to do
with the custom adapter. I'll get to that tomorrow.